fix error cd /workspaces/audio-ml-environment/data2vec2-spectrogram ; /usr/bin/env /opt/conda/bin/python /root/.vscode-server/extensions/ms-python.python-2023.4.1/pythonFiles/lib/python/debugpy/adapter/../../debugpy/launcher 59371 -- /workspace/src/fairseq/fairseq_cli/hydra_train.py --config-dir /workspaces/audio-ml-environment/data2vec2-spectrogram/examples/data2vec/config/v2 --config-name base_spectrogram_only_task common.user_dir=/workspaces/audio-ml-environment/data2vec2-spectrogram/examples/data2vec task.data=/workspaces/audio-ml-environment/data/librispeech/spectrograms/logspec/train-clean-100 distributed_training.distributed_world_size=1

+++ examples/data2vec/models/modalities/base.py
@@ -398,7 +398,8 @@ class ModalitySpecificEncoder(nn.Module):
                         indices=mask_seed.ids if mask_seed is not None else None,
                     )
 
-                    mask = torch.from_numpy(mask).to(device=x.device)
+                    mask = torch.from_numpy(mask).to(device=x.device, dtype=torch.int16)
+
                     if self.modality_cfg.inverse_mask:
                         mask = 1 - mask
                     mask_info = self.make_maskinfo(x, mask)
